chore(backend): reconcile and remove the diverged orphaned analytics.dto.ts duplicate
The root-level analytics.dto.ts was an unused, diverged duplicate of
dto/analytics.dto.ts. It turned out to hold five DTOs (GetAggregateBurnQueryDto,
TokenBurnSummaryDto, TopBurnerDto, BurnRateTrendPointDto,
AggregateBurnResponseDto) that analytics.controller.ts and analytics.service.ts
already import from dto/analytics.dto.ts but that file didn't export. Ported
those DTOs into dto/analytics.dto.ts and removed the orphaned root-level file.

Closes #1869
